feat: added a message to the interface "what would you like to make?"
