+++ gui/src/pages/creator/inputBox/pearIcon.tsx
@@ -1,0 +1,12 @@
+import { FC } from "react"
+
+export const PearIcon: FC<{className?: string}> = ({ className }) => {
+    return (
+        <svg width="19" height="19" viewBox="0 0 19 19" fill="none" xmlns="http://www.w3.org/2000/svg" className={className}>
+            <path fill-rule="evenodd" clip-rule="evenodd" d="M9.50003 5.04493C9.50267 5.04492 9.50531 5.04492 9.50795 5.04492V5.04493C11.7052 5.04624 11.7286 5.13712 12.1604 6.81366L12.1604 6.81367C12.2608 7.20342 12.3833 7.67886 12.5603 8.2588C13.096 10.0141 13.1977 10.2704 13.3148 10.5653C13.404 10.7902 13.5022 11.0375 13.8086 11.9896C14.4995 14.1357 13.6819 16.9317 9.50795 17.0679V17.0684C9.50531 17.0683 9.50267 17.0682 9.50004 17.0681C9.4974 17.0682 9.49476 17.0683 9.49212 17.0684V17.0679C5.31818 16.9317 4.50061 14.1357 5.19143 11.9896C5.49787 11.0375 5.59605 10.7902 5.68531 10.5653C5.80235 10.2704 5.90406 10.0141 6.43977 8.2588C6.61676 7.67886 6.73923 7.20341 6.83963 6.81366L6.83963 6.81366C7.2715 5.13712 7.29491 5.04624 9.49212 5.04493V5.04492C9.49476 5.04492 9.4974 5.04492 9.50003 5.04493Z" fill="black" />
+            <circle cx="11.6895" cy="9.9082" r="0.482422" fill="black" />
+            <path fill-rule="evenodd" clip-rule="evenodd" d="M10.6397 4.47268L10.6314 4.48005C10.6311 4.47728 10.6307 4.47411 10.6304 4.47055C10.6265 4.46966 10.6231 4.46885 10.6202 4.46815L10.6294 4.45994C10.6109 4.25962 10.5503 3.22199 11.1926 2.5163C11.847 1.79739 13.2512 1.89348 13.4799 1.9144L13.4836 1.91113L13.5122 1.9118C13.5122 1.9118 13.4884 3.54058 12.6251 4.15509C11.8479 4.70831 10.8342 4.51645 10.6397 4.47268Z" fill="black" />
+            <path d="M12.8438 12.8784C12.8438 14.0124 11.2852 14.8438 11.2852 14.8438C11.2852 14.8438 12.1758 14.0124 12.1758 12.8784C12.1758 11.7443 11.3742 11.2943 11.9531 11.089C12.532 10.8837 12.8438 11.7443 12.8438 12.8784Z" fill="black" />
+        </svg>
+    )
+}
